+++ example/index.js
@@ -2,7 +2,7 @@
 import "babel-polyfill";
 import "ol/ol.css";
 
-import {Style, Stroke, Fill} from "ol/style.js";
+import {Style, Stroke, Fill, Icon} from "ol/style.js";
 
 import * as mpapi from "../src";
 import abstractAPI from "../abstraction/api.js";
@@ -23,7 +23,7 @@ window.mpapi = {
 document.getElementById(portalConfig.target).innerHTML = "";
 //add a click-listener to button, that creates a 3D-map on click
 document.getElementById("enable").addEventListener("click", function (portalConfig) {
-    if(window.mpapi.map2D === null){
+    if (window.mpapi.map2D === null) {
         window.mpapi.map2D = window.mpapi.map;
     }
 
@@ -33,12 +33,12 @@ document.getElementById("enable").addEventListener("click", function (portalConf
     }
     else {
         const lib = portalConfig.cesiumLib ? portalConfig.cesiumLib : "https://geoportal-hamburg.de/mastercode/cesium/1_84/Cesium.js";
-        
-        load3DScript(lib, function Loaded3DCallback () {
+
+        load3DScript(lib, function Loaded3DCallback() {
             const settings3D = {
                 "map2D": window.mpapi.map2D
             },
-            map3D = abstractAPI.map.createMap(settings3D, {}, "3D");
+                map3D = abstractAPI.map.createMap(settings3D, "3D");
 
             window.mpapi.map = map3D;
             window.mpapi.map.setEnabled(true);
@@ -57,12 +57,12 @@ const hamburgServicesUrl = "http://geoportal-hamburg.de/lgv-config/services-inte
         ...portalConfig,
         layerConf: services
     },
-    map2D = abstractAPI.map.createMap(config, {}, "2D");
+    map2D = abstractAPI.map.createMap(config, "2D");
 
 services.push(localService);
 // */
 
-//* geojson styling function call to override default styling
+//* geojson styling function call to override default styling and special wfs styling
 mpapi.geojson.setCustomStyles({
     MultiPolygon: new Style({
         stroke: new Stroke({
@@ -74,12 +74,29 @@ mpapi.geojson.setCustomStyles({
         })
     })
 });
+
+const styleWfs = function (feature, resolution) {
+    const icon = new Style({
+        image: new Icon({
+            src: "https://img.icons8.com/windows/50/000000/building.png",
+            scale: 0.5,
+            opacity: 1
+        })
+    });
+    
+    return [icon];
+};
+services.forEach(entry => {
+    if (entry.typ === "WFS") {
+        entry["style"] = styleWfs;
+    }
+});
 // */
 
 //* SYNCHRONOUS EXAMPLE: layerConf is known
 window.mpapi.map = map2D;
 
-["2001","2002", "1002", "3001"].forEach(id => window.mpapi.map.addLayer(id));
+["2001", "2002", "1002", "3001"].forEach(id => window.mpapi.map.addLayer(id));
 
 
 // */
@@ -93,11 +110,11 @@ mpapi.rawLayerList.initializeLayerList(
             ...portalConfig,
             layerConf: conf,
             layers: null
-        });
+        }, "2D");
         ["6357", "6074"].forEach(
             id => window.mpapi.map.addLayer(id)
         );
-    }, "2D"
+    }
 );
 //*/
 
@@ -107,14 +124,13 @@ window.mpapi.map = abstractAPI.map.createMap(
     {...portalConfig, layerConf: hamburgServicesUrl, layers: [
         { id: "6357" },
         { id: "453", transparency: 50 }
-    ]},
+    ]}, "2D",
     {
         callback: () =>
             ["6074"].forEach(
                 id => window.mpapi.map.addLayer(id)
             )
-    },
-    "2D"
+    }
 );
 //*/