+++ jest.setup.js
@@ -1,3 +1,4 @@
 window.URL.createObjectURL = function () {
     // empty
 };
+require("regenerator-runtime/runtime");